fix: loadServeEnv in different runtime
