+++ packages/server/core/src/base/index.ts
@@ -7,6 +7,7 @@ import {
   HonoEnv,
 } from './types';
 import { createNodeServer } from './adapters/node';
+import { loadServerEnv } from './adapters/env';
 import { httpCallBack2HonoMid, connectMid2HonoMid } from './adapters/hono';
 import { registerMockHandlers } from './adapters/mock';
 
@@ -23,7 +24,7 @@ export * from './middlewares';
 
 export { httpCallBack2HonoMid, connectMid2HonoMid };
 export type { ServerNodeContext, ServerBaseOptions, Next, Middleware };
-export { registerMockHandlers };
+export { registerMockHandlers, loadServerEnv };
 export { createNodeServer, ServerBase };
 
 export async function createServerBase<E extends HonoEnv>(
